mindforgepy/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
SQLAlchemy==2.0.26
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
